[FE] feat: Overlay Component 위치를 Playbar 안쪽으로 이동
